No longer resize dice tray when it's full screen
This is to fix an issue with the dice tray window leaving full screen (F11) when the main window was entering full screen. If on a second monitor the dice tray should also enter full screen (F11) when the main window does. This can only be achieved manually on the same monitor currently.