Resize window fix for fullscreen dice tray
When the dice tray window is full screen (F11) with another window open on top of it it's inner height is actually 1 less than screen height. This was causing the full screen dice tray to resize unintentionally when behind another window and changing the main windows size. This behavior should be eliminated entirely now.